feat(ui): change member component to share component
