+++ src/ui/pages/NotificationDetails/components/MultisigMember/MultisigMember.types.ts
@@ -1,0 +1,14 @@
+enum MemberAcceptStatus {
+  Accepted,
+  Waiting,
+  Rejected,
+}
+
+interface MemberProps {
+  name: string;
+  status: MemberAcceptStatus;
+}
+
+export type { MemberProps };
+
+export { MemberAcceptStatus };
