+++ src/ui/pages/NotificationDetails/components/MultisigMember/index.ts
@@ -1,0 +1,2 @@
+export * from "./MultisigMember";
+export * from "./MultisigMember.types";
