+++ src/ui/pages/NotificationDetails/components/ReceiveCredential/ReceiveCredential.scss
@@ -19,7 +19,7 @@
   }
 
   .request-animation-center {
-    min-height: calc(100vh - (12.6rem + var(--ion-safe-area-top)));
+    min-height: calc(100vh - (13.5rem + var(--ion-safe-area-top)));
     height: 100%;
     display: flex;
     justify-content: center;
@@ -519,61 +519,6 @@
     .page-footer {
       animation-delay: 1s;
       animation: translateFooterDownAnimation 1s forwards;
-    }
-  }
-}
-
-.member {
-  --padding-start: 0;
-  --inner-padding-end: 0;
-  --padding-top: 0.75rem;
-  --padding-bottom: 0.75rem;
-  --min-height: auto;
-
-  .member-avatar {
-    width: 1.5rem;
-    height: 1.5rem;
-  }
-
-  .member-name {
-    width: 70%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-
-    &.md {
-      margin-inline-end: 0;
-    }
-  }
-
-  .status {
-    border-radius: 50%;
-    padding: 0.3125rem;
-    width: 0.875rem;
-    height: 0.875rem;
-
-    &.accepted {
-      background: var(--ion-color-primary-gradient);
-    }
-
-    &.rejected {
-      background: #ff7d7a;
-    }
-
-    &.waiting {
-      background: rgba(var(--ion-color-dark-grey-rgb), 0.5);
-    }
-  }
-
-  @media screen and (min-width: 250px) and (max-width: 370px) {
-    .status {
-      padding: 0.125rem;
-      width: 0.675rem;
-      height: 0.675rem;
-
-      &.md {
-        margin-inline-start: 0;
-      }
     }
   }
 }
